myFunctionProject/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pymupdf
pdf2image
pytesseract
aiohttp
numpy
pandas
azure-storage-blob
python-multipart
python-dotenv
azure-ai-formrecognizer

